requirements/requirements.txt
# local package
-e .

# hard dependencies
pip==21.3.1
torch==2.1.0 
--index-url https://download.pytorch.org/whl/cu118

transformers==4.40.1
#git+https://github.com/huggingface/transformers
datasets==2.19.0
peft==0.5.0
bitsandbytes==0.43.1
trl==0.7.1

# tensorboard
tensorboard

# external requirements
click
Sphinx
coverage
#awscli
flake8
python-dotenv>=0.5.1
loguru
